inline HqUserDetailsService implementation
